vectorplusplus/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
supabase==2.10.0
sentence-transformers==3.3.1
scikit-learn==1.5.2
playwright==1.49.0
PyGithub==2.5.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.10.3
numpy==1.26.4
docker==7.1.0
httpx==0.27.2
